feat: add plugin page view tracking and plugin web attribution
