+++ packages/analytics-browser/test/index.test.ts
@@ -1,6 +1,5 @@
 import {
   add,
-  AmplitudeBrowser,
   createInstance,
   flush,
   getDeviceId,
@@ -28,7 +27,6 @@ import {
 describe('index', () => {
   test('should expose apis', () => {
     expect(typeof add).toBe('function');
-    expect(typeof AmplitudeBrowser).toBe('function');
     expect(typeof createInstance).toBe('function');
     expect(typeof flush).toBe('function');
     expect(typeof groupIdentify).toBe('function');
