+++ packages/analytics-types/src/index.ts
@@ -22,9 +22,6 @@ export {
   InitOptions,
   NodeConfig,
   NodeOptions,
-  PageTrackingBrowserOptions,
-  PageTrackingReactNativeOptions,
-  PageTrackingFilter,
   ReactNativeConfig,
   ReactNativeOptions,
   ReactNativeTrackingOptions,
