--- packages/plugin-ma-tracking/src/index.ts
@@ -1,1 +1,0 @@
-export { MATrackingPlugin } from './ma-tracking';
