+++ packages/plugin-page-view-tracking/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@amplitude/plugin-ma-tracking",
+  "name": "@amplitude/plugin-page-view-tracking",
   "version": "0.0.0",
   "description": "",
   "author": "Amplitude Inc",
