+++ packages/plugin-page-view-tracking/src/index.ts
@@ -1,0 +1,2 @@
+export { PageViewTrackingPlugin } from './page-view-tracking';
+export * from './typings/page-view-tracking';
