+++ packages/plugin-page-view-tracking/src/typings/page-view-tracking.ts
@@ -1,0 +1,5 @@
+export interface PageTrackingBrowserOptions {
+  filter?: PageTrackingFilter;
+}
+
+export type PageTrackingFilter = 'onAttribution' | (() => boolean) | undefined;
