+++ packages/plugin-page-view-tracking/tsconfig.es5.json
