+++ packages/plugin-page-view-tracking/tsconfig.json
