+++ packages/plugin-web-attribution/src/index.ts
@@ -1,0 +1,1 @@
+export { WebAttributionPlugin } from './web-attribution';
